requirements.txt
alabaster==0.7.16
asgiref==3.7.2
attrs==23.2.0
Babel==2.14.0
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
commonmark==0.9.1
dj-database-url==2.1.0
Django==5.0.2
django-markdownx==4.0.7
django-on-heroku==1.1.2
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.27.1
gunicorn==21.2.0
idna==3.6
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
Pygments==2.17.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1